Move back to original AbacusMessage
