--- typescript/sdk/src/core/message/index.ts
@@ -1,2 +1,0 @@
-export { BaseMessage } from './base';
-export { AbacusMessage, AbacusStatus, MessageStatus } from './message';
